move code to main directory (remove src)
