+++ master/interfaces.go
