+++ slave/components/assembler/assembler_test.go
