+++ slave/components/interfaces.go
