+++ slave/interfaces.go
