+++ slave/slave.go
@@ -5,9 +5,9 @@ package slave
 import (
 	"fmt"
 
-	"github.com/genomagic/src/slave/components/parser"
-	"github.com/genomagic/src/slave/components/assembler"
-	"github.com/genomagic/src/slave/components"
+	"github.com/genomagic/slave/components"
+	"github.com/genomagic/slave/components/assembler"
+	"github.com/genomagic/slave/components/parser"
 )
 
 const (
