+++ slave/slave_test.go
